+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2022-01-27
+### Added
+- Support for React Icons
+- Animated Background gradient with noise
+- Contact icons
+- Layout component for the pages
+- Section component for the page section
+### Removed
+- The container component
+### Fixed
+- Github actions to run in node 12 environment as well
+
+
+
 ## [0.1.0] - 2022-01-20
 ### Added
 - Basic template to start development on using [superplate CLI](https://pankod.github.io/superplate/)